refactor: isolate shared clash config support
